requirements.txt
flask
gunicorn
tensorflow-cpu==2.11.0
numpy
pandas
scikit-learn